+++ crates/tauri-store-cli/assets/plugin-rs/plugin.rs
@@ -1,7 +1,13 @@
-use std::sync::Arc;
-use tauri::Runtime;
-use tauri_store::{Collection, StoreCollection};
+use tauri::{Runtime, State};
+use tauri_store::{Collection, CollectionMarker, StoreCollection};
 
 /// The __PASCAL_PLUGIN_TITLE__ plugin.
 #[derive(Collection)]
-pub struct __PASCAL_PLUGIN_TITLE__<R: Runtime>(pub(crate) Arc<StoreCollection<R>>);
+pub struct __PASCAL_PLUGIN_TITLE__<'a, R: Runtime>(
+  pub(crate) State<'a, StoreCollection<R, __PASCAL_PLUGIN_TITLE__Marker>>,
+);
+
+/// Marker for the __PASCAL_PLUGIN_TITLE__ plugin.
+pub struct __PASCAL_PLUGIN_TITLE__Marker;
+
+impl CollectionMarker for __PASCAL_PLUGIN_TITLE__Marker {}