+++ crates/tauri-store/src/collection/builder.rs
@@ -1,12 +1,15 @@
-use super::{OnLoadFn, StoreCollection, RESOURCE_ID};
+use super::marker::CollectionMarker;
+use super::{DefaultMarker, OnLoadFn, StoreCollection};
 use crate::collection::autosave::Autosave;
 use crate::error::Result;
 use crate::meta::Meta;
 use crate::store::{SaveStrategy, Store, StoreId};
+use crate::ManagerExt;
 use dashmap::DashMap;
 use std::collections::HashSet;
+use std::marker::PhantomData;
 use std::path::{Path, PathBuf};
-use std::sync::{Arc, Mutex};
+use std::sync::Mutex;
 use std::time::Duration;
 use tauri::{Manager, Runtime};
 
@@ -17,11 +20,15 @@ use tauri::plugin::TauriPlugin;
 use crate::migration::{Migration, MigrationContext, Migrator};
 
 /// Builder for the [`StoreCollection`](crate::collection::StoreCollection).
-pub struct StoreCollectionBuilder<R: Runtime> {
+pub struct StoreCollectionBuilder<R, C>
+where
+  R: Runtime,
+  C: CollectionMarker,
+{
   path: Option<PathBuf>,
   default_save_strategy: SaveStrategy,
   autosave: Option<Duration>,
-  on_load: Option<Box<OnLoadFn<R>>>,
+  on_load: Option<Box<OnLoadFn<R, C>>>,
   pretty: bool,
   save_denylist: Option<HashSet<StoreId>>,
   sync_denylist: Option<HashSet<StoreId>>,
@@ -30,7 +37,11 @@ pub struct StoreCollectionBuilder<R: Runtime> {
   migrator: Migrator,
 }
 
-impl<R: Runtime> StoreCollectionBuilder<R> {
+impl<R, C> StoreCollectionBuilder<R, C>
+where
+  R: Runtime,
+  C: CollectionMarker,
+{
   /// Creates a new builder instance with default values.
   pub fn new() -> Self {
     Self::default()
@@ -54,7 +65,7 @@ impl<R: Runtime> StoreCollectionBuilder<R> {
   #[must_use]
   pub fn on_load<F>(mut self, f: F) -> Self
   where
-    F: Fn(&Store<R>) -> Result<()> + Send + Sync + 'static,
+    F: Fn(&Store<R, C>) -> Result<()> + Send + Sync + 'static,
   {
     self.on_load = Some(Box::new(f));
     self
@@ -149,34 +160,23 @@ impl<R: Runtime> StoreCollectionBuilder<R> {
     self
   }
 
-  /// Initializes the plugin with a [`StoreCollection`](crate::collection::StoreCollection).
-  ///
-  /// # Panics
-  ///
-  /// Panics if a store collection is already initialized.
-  #[cfg(feature = "plugin")]
-  pub fn build_plugin(self) -> TauriPlugin<R> {
-    crate::plugin::build(self)
-  }
-
   /// Builds the [`StoreCollection`](crate::collection::StoreCollection).
   ///
   /// # Panics
   ///
   /// Panics if a store collection is already initialized.
   #[doc(hidden)]
-  pub fn build<M>(mut self, app: &M, name: &str) -> Result<Arc<StoreCollection<R>>>
+  pub fn build<M>(mut self, app: &M, plugin_name: &str) -> Result<()>
   where
     M: Manager<R>,
   {
-    assert!(
-      RESOURCE_ID.get().is_none(),
+    let app = app.app_handle();
+    debug_assert!(
+      app.try_state::<StoreCollection<R, C>>().is_none(),
       "store collection is already initialized"
     );
 
-    let app = app.app_handle();
-    let meta = Meta::read(app, name)?;
-
+    let meta = Meta::read(app, plugin_name)?;
     let path = meta
       .inner
       .path
@@ -186,7 +186,7 @@ impl<R: Runtime> StoreCollectionBuilder<R> {
           .path()
           .app_data_dir()
           .expect("failed to resolve app data dir")
-          .join(name)
+          .join(plugin_name)
       });
 
     #[cfg(feature = "unstable-migration")]
@@ -197,9 +197,9 @@ impl<R: Runtime> StoreCollectionBuilder<R> {
     self.save_denylist = self.save_denylist.filter(|it| !it.is_empty());
     self.sync_denylist = self.sync_denylist.filter(|it| !it.is_empty());
 
-    let collection = Arc::new(StoreCollection::<R> {
+    app.manage(StoreCollection::<R, C> {
       app: app.clone(),
-      name: Box::from(name),
+      name: Box::from(plugin_name),
       path: Mutex::new(path),
       stores: DashMap::new(),
       on_load: self.on_load,
@@ -208,24 +208,43 @@ impl<R: Runtime> StoreCollectionBuilder<R> {
       save_denylist: self.save_denylist,
       sync_denylist: self.sync_denylist,
       pretty: self.pretty,
+      phantom: PhantomData,
 
       #[cfg(feature = "unstable-migration")]
       migrator: Mutex::new(self.migrator),
     });
 
-    let rid = app
-      .resources_table()
-      .add_arc(Arc::clone(&collection));
+    app
+      .store_collection_with_marker::<C>()
+      .autosave
+      .lock()
+      .unwrap()
+      .start::<R, C>(app);
 
-    let _ = RESOURCE_ID.set(rid);
-
-    collection.autosave.lock().unwrap().start(app);
-
-    Ok(collection)
+    Ok(())
   }
 }
 
-impl<R: Runtime> Default for StoreCollectionBuilder<R> {
+impl<R> StoreCollectionBuilder<R, DefaultMarker>
+where
+  R: Runtime,
+{
+  /// Initializes the plugin with a [`StoreCollection`](crate::collection::StoreCollection).
+  ///
+  /// # Panics
+  ///
+  /// Panics if a store collection is already initialized.
+  #[cfg(feature = "plugin")]
+  pub fn build_plugin(self) -> TauriPlugin<R> {
+    crate::plugin::build(self)
+  }
+}
+
+impl<R, C> Default for StoreCollectionBuilder<R, C>
+where
+  R: Runtime,
+  C: CollectionMarker,
+{
   fn default() -> Self {
     Self {
       path: None,