+++ crates/tauri-store/src/collection/marker.rs
@@ -1,0 +1,7 @@
+/// A marker trait used to enable plugin interoperability.
+pub trait CollectionMarker: Send + Sync + 'static {}
+
+/// Default collection marker.
+pub struct DefaultMarker;
+
+impl CollectionMarker for DefaultMarker {}