+++ examples/playground/src-tauri/src/lib.rs
@@ -1,6 +1,5 @@
-use std::time::Duration;
 use tauri::{AppHandle, WebviewUrl, WebviewWindowBuilder};
-use tauri_plugin_pinia::{BoxResult, SaveStrategy};
+use tauri_plugin_pinia::BoxResult;
 
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
@@ -8,15 +7,11 @@ pub fn run() {
     .plugin(tauri_plugin_process::init())
     .plugin(tauri_plugin_shell::init())
     .plugin(tauri_plugin_window_state::Builder::new().build())
-    .plugin(
-      tauri_plugin_pinia::Builder::new()
-        .autosave(Duration::from_secs(60))
-        .default_save_strategy(SaveStrategy::throttle_secs(3))
-        .pretty(true)
-        .save_denylist(["dont-save-1", "dont-save-2"])
-        .sync_denylist(["dont-sync-1", "dont-sync-2"])
-        .build(),
-    )
+    .plugin(tauri_plugin_pinia::init())
+    .plugin(tauri_plugin_svelte::init())
+    .plugin(tauri_plugin_valtio::init())
+    .plugin(tauri_plugin_vue::init())
+    .plugin(tauri_plugin_zustand::init())
     .setup(|app| open_window(app.handle()))
     .invoke_handler(tauri::generate_handler![])
     .run(tauri::generate_context!())