+++ docs/src/plugin-valtio/changelog.md
@@ -12,6 +12,12 @@ outline: 2
 
 _Nothing yet._
 
+## 2.5.0
+
+### Dependencies
+
+- Update `tauri-store` to `0.11.0`.
+
 ## 2.4.0
 
 ### Features